# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Todostock API",
        "description": "Documentación oficial de los endpoints de la API de Todostock para Todotek S.A. Incluye autenticación por Sanctum, gestión de inventarios y facturación transaccional.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Servidor Local de Desarrollo (Docker Sail)"
        }
    ],
    "paths": {
...
        "/invoices": {
            "get": {
                "tags": [
                    "Facturación"
                ],
                "summary": "Obtener listado de facturas paginado",
                "description": "Retorna una lista paginada de las facturas emitidas con la información básica del cliente. Permite buscar por número de factura, nombre o identificación del cliente.",
                "operationId": "getInvoicesList",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Término de búsqueda (número de factura, nombre o identificación del cliente)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página para la paginación del servidor",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de facturas obtenido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "client_id": {
                                                        "type": "integer",
                                                        "example": 3
                                                    },
                                                    "invoice_number": {
                                                        "type": "string",
                                                        "example": "FAC-000001"
                                                    },
                                                    "subtotal": {
                                                        "type": "string",
                                                        "example": "150.00"
                                                    },
                                                    "tax_total": {
                                                        "type": "string",
                                                        "example": "22.50"
                                                    },
                                                    "total": {
                                                        "type": "string",
                                                        "example": "172.50"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2026-05-19T00:00:00.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2026-05-19T00:00:00.000000Z"
                                                    },
                                                    "client": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 3
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Kevin Cepeda"
                                                            },
                                                            "identification": {
                                                                "type": "string",
                                                                "example": "0955555555"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 15
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Facturación"
                ],
                "summary": "Crear una nueva factura (Venta transaccional)",
                "description": "Procesa la venta de productos bajo una transacción de base de datos. Valida la disponibilidad de stock, calcula subtotales e impuestos en tiempo real y decrementa el inventario de forma automática.",
                "operationId": "storeInvoice",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "client_id",
                                    "items"
                                ],
                                "properties": {
                                    "client_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "items": {
                                        "description": "Listado de productos y cantidades a facturar",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Factura procesada y stock actualizado con éxito",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Factura procesada y stock actualizado con éxito."
                                        },
                                        "invoice": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 16
                                                },
                                                "client_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "invoice_number": {
                                                    "type": "string",
                                                    "example": "FAC-000016"
                                                },
                                                "subtotal": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 500
                                                },
                                                "tax_total": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 75
                                                },
                                                "total": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 575
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2026-05-19T00:00:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2026-05-19T00:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de negocio o validación (Ej: Stock insuficiente para el producto)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Stock insuficiente para el producto: Producto Ejemplo. Disponibles: 1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Introduce el token 'access_token' recibido en el login con el formato: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}